BoardSight/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
redis==5.2.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
stockfish==3.28.0
httpx==0.28.1
python-chess==1.999
pydantic==2.10.4
